@@ -4,32 +4,26 @@
   "private": true,
   "dependencies": {
     "@aspnet/signalr": "^1.1.4",
-    "@craco/craco": "^5.2.1",
-    "@types/classnames": "^2.2.8",
-    "@types/moment": "^2.13.0",
-    "@types/moment-timezone": "^0.5.12",
-    "@types/react-document-title": "^2.0.3",
-    "@types/react-loadable": "^5.5.1",
-    "@types/recharts": "^1.1.16",
-    "abp-web-resources": "^3.8.5",
-    "antd": "^3.19.2",
+    "@craco/craco": "^5.6.2",
+    "abp-web-resources": "^4.1.0",
+    "antd": "^3.26.0",
     "axios": "^0.19.0",
     "classnames": "^2.2.6",
-    "craco-antd": "^1.11.0",
+    "craco-antd": "^1.14.1",
     "famfamfam-flags": "^1.0.0",
-    "lodash": "^4.17.11",
-    "mobx": "^5.10.1",
-    "mobx-react": "^6.0.3",
+    "lodash": "^4.17.15",
+    "mobx": "^5.15.0",
+    "mobx-react": "^6.1.4",
     "moment": "^2.24.0",
-    "moment-timezone": "^0.5.25",
-    "query-string": "^6.7.0",
-    "react": "^16.8.6",
+    "moment-timezone": "^0.5.27",
+    "query-string": "^6.9.0",
+    "react": "^16.12.0",
     "react-document-title": "^2.0.3",
-    "react-dom": "^16.8.6",
+    "react-dom": "^16.12.0",
     "react-loadable": "^5.5.0",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "^3.0.1",
-    "recharts": "^1.6.2"
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "^3.2.0",
+    "recharts": "^1.8.5"
   },
   "scripts": {
     "start": "craco start",
@@ -38,14 +32,20 @@
     "eject": "react-scripts eject"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.13",
-    "@types/node": "^12.0.7",
-    "@types/react": "^16.8.19",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-router-dom": "^4.3.3",
-    "copy-webpack-plugin": "^5.0.3",
-    "ts-import-plugin": "^1.5.5",
-    "typescript": "^3.5.1"
+    "@types/classnames": "^2.2.9",
+    "@types/jest": "^24.0.23",
+    "@types/moment": "^2.13.0",
+    "@types/moment-timezone": "^0.5.12",
+    "@types/node": "^12.12.14",
+    "@types/react": "^16.9.13",
+    "@types/react-document-title": "^2.0.3",
+    "@types/react-dom": "^16.9.4",
+    "@types/react-loadable": "^5.5.2",
+    "@types/react-router-dom": "^5.1.3",
+    "@types/recharts": "^1.8.3",
+    "copy-webpack-plugin": "^5.0.5",
+    "ts-import-plugin": "^1.6.1",
+    "typescript": "^3.7.2"
   },
   "browserslist": {
     "production": [
